fix(assets): 🍱 remove deprecated svg style
"enable-background":
Deprecated: This feature is no longer recommended.
Though some browsers might still support it,
it may have already been removed from the relevant web standards,
may be in the process of being dropped, or may only be kept for compatibility purposes.
Avoid using it, and update existing code if possible;
Be aware that this feature may cease to work at any time.